@@ -1,13 +1,14 @@
 apply plugin: 'com.android.application'
 
 repositories{
+    google()
     jcenter()
 }
 
 
 android {
-    compileSdkVersion 23
-    buildToolsVersion "22.0.1"
+    compileSdkVersion 29
+    buildToolsVersion "29.0.2"
 
     defaultConfig {
         applicationId "com.sample.jsonstoreandroid"
@@ -31,13 +32,13 @@ android {
 }
 
 dependencies {
-    compile fileTree(dir: 'libs', include: ['*.jar'])
-    testCompile 'junit:junit:4.12'
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
+    testImplementation  'junit:junit:4.12'
 
-    compile 'com.android.support:appcompat-v7:23.1.1'
-    compile 'com.android.support:design:23.1.1'
+    implementation 'androidx.appcompat:appcompat:1.0.0'
+    implementation 'com.google.android.material:material:1.0.0'
 
-    compile 'com.ibm.mobile.foundation:ibmmobilefirstplatformfoundation:8.0.+'
-    compile 'com.ibm.mobile.foundation:ibmmobilefirstplatformfoundationjsonstore:8.0.+'
+    implementation 'com.ibm.mobile.foundation:ibmmobilefirstplatformfoundation:8.0.+'
+    implementation 'com.ibm.mobile.foundation:ibmmobilefirstplatformfoundationjsonstore:8.0.+'
 }
 
